+++ lib/src/universal/extensions/podcast/podcast_item_parser.dart
@@ -17,18 +17,31 @@ class PodcastItemParser implements ItemExtensionParser {
   void parseItem(UniversalFeed feed, Item item, XmlElement element) {
     final pi = item.podcast ?? PodcastItem();
 
-    element.forEachElementXml(
-      'txt',
-      (value) {
-        pi.txts.add(
-          PodcastTxt(
-            purpose: value.getAttribute('purpose'),
-            value: value.innerText.trim(),
-          ),
-        );
-      },
-      ns: namespaceUrl,
-    );
+    element
+      ..forEachElementXml(
+        'txt',
+        (value) {
+          pi.txts.add(
+            PodcastTxt(
+              purpose: value.getAttribute('purpose'),
+              value: value.innerText.trim(),
+            ),
+          );
+        },
+        ns: namespaceUrl,
+      )
+      ..forEachElementXml(
+        'chapters',
+        (value) {
+          final url = value.getAttribute('url')?.trim();
+          final type = value.getAttribute('type')?.trim();
+          pi.chapters = PodcastChapters(
+            url: (url == null || url.isEmpty) ? null : url,
+            type: (type == null || type.isEmpty) ? null : type,
+          );
+        },
+        ns: namespaceUrl,
+      );
 
     item.podcast = pi;
   }